[Fix] Fix 32-bit integer problems on Big (and Mixed) Endian hosts.
If we write a value to int64value and read it from int32value, then it will
always be 0. This is because the int32value is always read if the upper 32
bits of the int64value is zero. Since a Big Endian platform reads the most
significant bits/bytes first, it'll read the upper 32 bits, not the lower
32 bits.

Signed-off-by: Jens Bauer <jens-github@gpio.dk>